# docs/reference/openapi.json
{
  "components": {
    "schemas": {
      "CreateScanRequest": {
        "properties": {
          "ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ref"
          },
          "repo_url": {
            "format": "uri",
            "maxLength": 2083,
            "minLength": 1,
            "title": "Repo Url",
            "type": "string"
          }
        },
        "required": [
          "repo_url"
        ],
        "title": "CreateScanRequest",
        "type": "object"
      },
      "CreateScanResponse": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "scan_id": {
            "title": "Scan Id",
            "type": "integer"
          },
          "status": {
            "const": "queued",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "scan_id",
          "status",
          "job_id"
        ],
        "title": "CreateScanResponse",
        "type": "object"
      },
      "FindingItemResponse": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "file_path": {
            "title": "File Path",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "line_end": {
            "title": "Line End",
            "type": "integer"
          },
          "line_start": {
            "title": "Line Start",
            "type": "integer"
          },
          "recommendation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation"
          },
          "rule_id": {
            "title": "Rule Id",
            "type": "string"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "rule_id",
          "file_path",
          "line_start",
          "line_end",
          "severity",
          "confidence",
          "recommendation"
        ],
        "title": "FindingItemResponse",
        "type": "object"
      },
      "FindingListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FindingItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "scan_id": {
            "title": "Scan Id",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "version": {
            "const": "v1",
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "scan_id",
          "total",
          "limit",
          "offset",
          "items"
        ],
        "title": "FindingListResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ScanDetailResponse": {
        "properties": {
          "commit_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Sha"
          },
          "critical_count": {
            "title": "Critical Count",
            "type": "integer"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "high_count": {
            "title": "High Count",
            "type": "integer"
          },
          "low_count": {
            "title": "Low Count",
            "type": "integer"
          },
          "medium_count": {
            "title": "Medium Count",
            "type": "integer"
          },
          "repo_url": {
            "title": "Repo Url",
            "type": "string"
          },
          "scan_id": {
            "title": "Scan Id",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "total_findings": {
            "title": "Total Findings",
            "type": "integer"
          }
        },
        "required": [
          "scan_id",
          "repo_url",
          "commit_sha",
          "status",
          "started_at",
          "finished_at",
          "total_findings",
          "critical_count",
          "high_count",
          "medium_count",
          "low_count"
        ],
        "title": "ScanDetailResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Repo Secret & Dependency Auditor API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/scans": {
      "post": {
        "description": "Create a new security scan for a repository.\n\n**Security Notes:**\n- Rate limited to 10 requests per minute per IP.\n- Git URL validated against whitelist (github.com, gitlab.com, etc.).\n- Git ref validated to prevent malicious input.\n- All logging is sanitized to hide secrets.",
        "operationId": "create_scan_scans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateScanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Scan",
        "tags": [
          "scans"
        ]
      }
    },
    "/scans/{scan_id}": {
      "get": {
        "description": "Get details of a scan.",
        "operationId": "get_scan_scans__scan_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scan",
        "tags": [
          "scans"
        ]
      }
    },
    "/scans/{scan_id}/findings": {
      "get": {
        "description": "Get findings for a scan with optional filtering.\n\n**Security Notes:**\n- Rate limited to 30 requests per minute per IP.\n- Severity and type filters validated against allowed values.\n- Pagination enforced to prevent large result exfiltration.",
        "operationId": "get_scan_findings_scans__scan_id__findings_get",
        "parameters": [
          {
            "in": "path",
            "name": "scan_id",
            "required": true,
            "schema": {
              "title": "Scan Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Severity"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scan Findings",
        "tags": [
          "scans"
        ]
      }
    }
  }
}